+++ crates/prelude/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Patch
 
+- Fix clippy lints
 - Update dependencies
 
 ## 0.7.0
@@ -191,4 +192,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 1 -->
+<!-- Increment to skip CHANGELOG.md test: 2 -->